backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0
opencv-python>=4.9
numpy>=1.26
ultralytics>=8.2.18
torch>=2.2
torchvision>=0.17
lapx>=0.5.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
reportlab>=4.1
openpyxl>=3.1
